health_clinic/backend/requirements.txt
FastAPI
uvicorn
sqlalchemy
asyncpg
pydantic
python-jose
passlib[bcrypt]
pytest
httpx
jinja2